refactor: Short name for workspace config
